Rename client to hmr-client
