+++ crates/turbopack-ecmascript-runtime/js/package.json
@@ -7,7 +7,7 @@
   "scripts": {
     "check": "run-p check:*",
     "check:nodejs": "tsc -p src/nodejs",
-    "check:browser-client": "tsc -p src/browser/client",
+    "check:browser-client": "tsc -p src/browser/dev/hmr-client",
     "check:browser-runtime-base": "tsc -p src/browser/runtime/base",
     "check:browser-runtime-dom": "tsc -p src/browser/runtime/dom",
     "check:browser-runtime-edge": "tsc -p src/browser/runtime/edge"
